Add a functor for hashing the links
The `link_type` in `DirectedGraph` is a tuple containing its two sides
of the link so it requires a custom hash function. The hash function
defined in this commit computes the hash values for each side. Then, it
XORs these two values to get the final hash value for the link.