Fix default name-to-id callback function
First, we have to use `std::stoll` instead of `std::stoi` to convert
the name string to integer. Because the ID field is usually a 64-bit
integer (`long long int`) rather than `int`.

Second, we need a wrapper function because the `std::stoll` is a
template function and we need to pass a fully-qualified identifier as
input parameter.