Switch to sparse hashmap
The reason behind choosing sparsehash is its memory efficiency. So, it
make sense a lot to use `sparse_hash_map` instead of `dense_hash_map`.
As a result, there is no need to call `set_empty_key` for
`sparse_hash_map`.

Apart from that, the `init_rank_map` belongs to the `GraphTrait` instead
of `DirectedGraphTraits`. So, it has been moved to the proper class.