Consider node IDs as handles
Node IDs are no longer exposed in the API. They will be chosen internally such
that they can be used to efficiently access nodes and data associated with them.
They are like "handles" in xg graphs. Input graphs could only provide node names
not IDs. The IDs in a vg graph are considered as name by converting them to
string if `name` field is empty (most of the cases). Otherwise, the parser
embeds the ID into the name string such that it can be extracted again if
needed.