Use `const_reference` of the container in iterator class
The constant reference type usually should be `value_type const&`, unless the
container does not return any reference for the element access, such as
`StringSet` or `NodeProperty` classes. So, instead of defining to return by
value by default in the iterator class, we re-use the `const_reference` type of
the container as `const_reference` type of the iterator and let the container be
responsible for defining this type. This has an advantage that some unnecessary
copies can be avoided when the container returns a reference to the element.