Fix the issue of the recursive type definitions
The `const_reference` type is defined based on the `const_reference` type of the
container. So, we cannot define a container with a `const_reference` depending
on `const_reference` type of its iterator.

Apart from that, since the proxy container returns the element by value (not
reference) the const reference type would be `value_type const`.