Drop checking node ID, rank, and link type in API methods
The methods assume that the provided node ID, node rank, and link type is valid.
If the node ID, node rank, or link type is not valid, the behaviour of the
methods are undefined. A set of methods are defined to check these values before
calling these methods.

This decision is made because most of these values are provided internally and
there are rare cases that these values are provided from outside. For example,
node IDs are assigned by the `DirectedGraph` classes, and the values are usually
provided by other methods in the class; for example `for_each_node`. So,
checking seems unnecessary in those cases.