Fix side type for Directed graph
The `Side` class depends on the template parameter of outer class. This means
that the side types are different for different graph with different template
parameter. To fix this, this commit replaces the `Side` class with an
independent type, here `std::tuple< id_type >`. In addition, it would be more
compatible with the side type of the Bidirected graph.