Improve the build script
This commit exports `gum::gum` as an interface library and defines the
library dependencies. The CMake modules are also updated to export the
targets for each dependencies. The `CMakeLists.txt` for the test module
is also updated to use `gum::gum` exported target. It also ensures that
`ext/catch2` submodule is fetched before trying to include it.