Change policy of assigning node IDs
Node IDs can be assigned externally for `Dynamic` graphs. So, the coordinate
system of a graph can be preserved when it is loaded into a `Dynamic` graph.
Those IDs are also stored in `Succinct` graphs, so a local ID can be translated
to the original IDs. But translating back is not efficient with graph API, and
requires an auxiliary data structure.