Introduce `Coordinate` class
Coordinate class is an auxiliary data structure which helps a graph to maintain
a coordinate system when directly embedding that coordinate system is not
possible. For example, embedding a coordinate system in `Succinct` graphs (since
node IDs are immutable) or those system that the node IDs are not integer (e.g.
GFA with non-numerical sequence segment name).