Fix the issue with `add_nodes` method
The `add_nodes` methods has been added to add a batch of node without specifying
their IDs. One of the overload of this method was unnecessary (when the node IDs
are known). Also, the `SeqGraph` did not override this method, which is added in
this commit. The overridden method adds corresponding node objects to the node
property member variable as well.