Improve `resize` method of `StringSet` class
Instead creating a new bigger bit vector and transfer the original data to it,
resize the existing bit vector and zero-initialise the expanded space. This
would be more efficient for pushing back a new element.

In addition, it avoids resizing when the requested size is equal to the current
size.