Fix a major bug in `StringSet`
Apparently, the bug that commit 4c08ca1 tries to fix is caused because of the
wrong assumption that all elements are non-empty strings. Therefore, that commit
does not actually fix the issue. Now, `StringSet` puts a delimiter at the end of
each element when they are concatenated to the main container. The bit vector is
set at the delimiter position of each element instead of the first bit.