Refactor GFA utils to conform with generic IO API
In order to decouple the library from external dependencies, there are
two sets of API interface functions: 1) template interface functions in
which data format is specified by a format specifier tag (duck typing).
2) interface functions overloads with external data types. GFA utils
module did not have the first set of interface functions which have been
implemented in this commit.