Make `Stoid` coordinate more robust
When segments IDs are not integers in a GFA file, one should use custom
coordinate to enumerate the nodes (e.g. `Sparse` coordinate will do).
In practice, node names can be prefixed with non-digit characters (e.g.
"s203"). `Stoid`, now, first try to convert node name to an integer id
and ignore any non-digit suffixes. If it fails, it tries again to check
if the name string ends with an integer. If yes, it is considered as
node id.