Better `RandomAccessProxyContainer` implementation
The underlying container might return a reference to its element when
indexed by `operator[]`. The previous implementation returns a copy if
`TValue` is not a lvalue-reference.

In addition, `function_type` should had been compatible with the
`TValue` provided as a template parameter. Also, the class would only
provide a constant proxy container (return const reference to the
elements).

Another issue with the implementation was use of `std::function`. Any
captured identifiers in the input lambda would be have been stored in
heap by `std::function` which makes it more expense than lambda.

The new implementation gets function type as template parameter instead
of value type. The latter will be deduced internally. This provides
easier syntax for construction of a proxy container too as all template
parameters can be deduced from the ctor arguments. Also, it allows
function type to be more flexible in terms of return value and input
parameters.

The new proxy container is not constant by default and its const-ness is
determined by the template type parameter `TContainer`. Also adding
const qualifier to the `RandomAccessProxyContainer` also preserves
const-ness of the underlying container by providing different access
methods for const and non-const instances.