Fix proxy containers for Dynamic node properties
This commit changes `SequenceProxyContainer` and `NameProxyContainer` to
be aligned with the new `RandomAccessProxyContainer` implementation and
brings the same rationale to these classes. Both proxy containers for
node properties (which are only used in Dynamic SeqGraph) are now can
work with non-const containers too. However, the old functionality; i.e.
exposing 'const' references to node properties; is preserved.