Fix a bug in DFS traverse function
The starting node of a region can have non-zero in-degree on its end
side. This commit instead searches for nodes with zero in-degree for
their start side. This choice affects drastically the topological sort
order and preserves the order of node in accordance with their
corresponding regions in the graph.